move complex store creation and add test for NavPanel
