+++ forest_lite/client/src/Nav.js
@@ -43,7 +43,7 @@ const pointURL = (baseURL, datasetID, variable, dimension) => {
 }
 
 
-const NavPanel = ({ baseURL, datasetName, dataVar }) => {
+export const NavPanel = ({ baseURL, datasetName, dataVar }) => {
     // Dimensions
     const [ dimension, setDimension ] = useState(null)
     const [ dimensions, setDimensions ] = useState([])
@@ -67,7 +67,6 @@ const NavPanel = ({ baseURL, datasetName, dataVar }) => {
             const variables = getDataVars(datasets[index])
             if ( variables.indexOf(dataVar) !== -1 ) {
                 const description = datasets[index].description
-                console.log(description)
                 const dims = description.data_vars[dataVar].dims
                 const dimensions = dims.filter(dim => {
                     return ["latitude", "longitude"].indexOf(dim) === -1
