+++ forest_lite/client/src/index.js
@@ -3,17 +3,8 @@ import React from "react"
 import ReactDOM from "react-dom"
 import { Provider } from "react-redux"
 import App from "./App.js"
-import { rootReducer } from "./reducers.js"
-import { toolMiddleware } from "./middlewares.js"
-import { animationMiddleware } from "./animation-middleware.js"
-import { colorPaletteMiddleware } from "./colorpalette-middleware.js"
-import { timeMiddleware } from "./time-middleware.js"
-import { zoomMiddleware } from "./zoom-middleware.js"
+import { createStore } from "./create-store.js"
 import {
-    SET_DATASETS,
-} from "./action-types.js"
-import {
-    set_dataset,
     set_datasets,
     set_limits,
     next_time_index,
@@ -21,38 +12,9 @@ import {
     fetch_image_success
 } from "./actions.js"
 
-// ReduxJS
-
-// Middlewares
-let logActionMiddleware = store => next => action => {
-    console.log(action)
-    next(action)
-}
-
-let datasetsMiddleware = store => next => action => {
-    next(action)
-    if (action.type == SET_DATASETS) {
-        const { dataset } = store.getState()
-        if (typeof dataset !== "undefined") {
-            let dataset_id = action.payload[0].id
-            next(set_dataset(dataset_id))
-        }
-    }
-    return
-}
-
 
 window.main = function(baseURL) {
-    let store = Redux.createStore(rootReducer,
-                                  Redux.applyMiddleware(
-                                      logActionMiddleware,
-                                      toolMiddleware,
-                                      animationMiddleware,
-                                      colorPaletteMiddleware,
-                                      timeMiddleware,
-                                      datasetsMiddleware,
-                                      zoomMiddleware,
-                                  ))
+    let store = createStore()
     // store.subscribe(() => { console.log(store.getState()) })
 
     // Use React to manage components
